refactor: move run into the same module as executor, but now no longer as a static class method to avoid confusion
